feat: implement Standalone Cart Page
Create a cart module. Implement a standalone cart page. Fix useCart linters and cartGetters linters.
Move an useCart composable, a cartGetters and cart store to the cart module.

re #340
