+++ packages/theme/components/BottomNavigation.vue
@@ -62,7 +62,7 @@
       <SfBottomNavigationItem
         :label="$t('Cart')"
         data-testid="bottom-navigation-cart"
-        @click="toggleCartSidebar"
+        @click="handleCartClick"
       >
         <template #icon>
           <SfCircleIcon aria-label="Go to cart">
@@ -100,7 +100,6 @@ export default defineComponent({
   },
   setup() {
     const {
-      toggleCartSidebar,
       toggleWishlistSidebar,
       toggleLoginModal,
       toggleMobileMenu,
@@ -126,6 +125,10 @@ export default defineComponent({
       }
     };
 
+    const handleCartClick = async () => {
+      await router.push(app.localeRoute({ name: 'cart' }));
+    };
+
     const loadCategoryMenu = async () => {
       const categories = useCategoryStore();
       if (categories.categories === null) {
@@ -138,11 +141,11 @@ export default defineComponent({
       isAuthenticated,
       isMobileMenuOpen,
       toggleWishlistSidebar,
-      toggleCartSidebar,
       toggleMobileMenu,
       loadCategoryMenu,
       handleAccountClick,
       handleHomeClick,
+      handleCartClick,
     };
   },
 });
