+++ packages/theme/components/CouponCode.vue
@@ -35,8 +35,8 @@ import {
   ref,
   defineComponent,
 } from '@nuxtjs/composition-api';
-import cartGetters from '~/modules/checkout/getters/cartGetters';
-import { useCart } from '~/modules/checkout/composables/useCart';
+import { getAppliedCoupon } from '~/modules/cart/getters/cartGetters';
+import { useCart } from '~/modules/cart/composables/useCart';
 
 export default defineComponent({
   name: 'CouponCode',
@@ -51,7 +51,7 @@ export default defineComponent({
 
     const couponCodeUserInput = ref('');
 
-    const couponCodeAppliedToCart = computed(() => cartGetters.getAppliedCoupon(cart.value)?.code);
+    const couponCodeAppliedToCart = computed(() => getAppliedCoupon(cart.value)?.code);
     const isCouponCodeApplied = computed(() => couponCodeAppliedToCart.value !== undefined);
     const hasAnyError = computed(() => Object.values(error.value).some((value) => value !== null));
     const applyCouponMsg = computed<Error>(() => error.value.applyCoupon || error.value.removeCoupon || { message: '', name: 'apply-coupon' });
