+++ packages/theme/composables/index.ts
@@ -9,7 +9,7 @@
 export * from './useApi';
 export * from '../modules/customer/composables/useAddresses';
 export * from '../modules/checkout/composables/useBilling';
-export * from '../modules/checkout/composables/useCart';
+export * from '../modules/cart/composables/useCart';
 export * from '../modules/catalog/category/composables/useCategory';
 export * from '../modules/catalog/category/composables/useCategorySearch';
 export * from './useConfig';
