+++ packages/theme/modules/cart/composables/useCart/commands/removeItemCommand.ts
