+++ packages/theme/modules/cart/getters/cartGetters.ts
@@ -12,8 +12,8 @@ import {
   ConfigurableCartItem,
   CartItemInterface,
 } from '~/modules/GraphQL/types';
-import { CartGetters as CartGettersBase, CartDiscount, Coupon } from './types';
 import { getName, getSlug as getSlugGetter, getProductThumbnailImage } from '~/modules/catalog/product/getters/productGetters';
+import { CartGetters as CartGettersBase, CartDiscount, Coupon } from './types';
 
 export const getItems = (cart: Cart): CartItemInterface[] => {
   if (!cart || !cart.items) {
