+++ packages/theme/modules/catalog/product/components/product-types/bundle/BundleProduct.vue
@@ -148,7 +148,7 @@ import {
   getAverageRating,
 } from '~/modules/review/getters/reviewGetters';
 
-import { useCart } from '~/modules/checkout/composables/useCart';
+import { useCart } from '~/modules/cart/composables/useCart';
 import useWishlist from '~/modules/wishlist/composables/useWishlist';
 import BundleProductSelector from '~/modules/catalog/product/components/product-types/bundle/BundleProductSelector.vue';
 import SvgImage from '~/components/General/SvgImage.vue';
