+++ packages/theme/modules/catalog/product/components/product-types/bundle/BundleProductSelector.vue
@@ -102,7 +102,7 @@ import { bundleProductInitialSelector } from '~/modules/catalog/product/helpers/
 import type { WithTypename } from '~/modules/catalog/product/types';
 import type { BundleProduct } from '~/modules/GraphQL/types';
 
-import useCart from '~/modules/checkout/composables/useCart';
+import { useCart } from '~/modules/cart/composables/useCart';
 
 export default defineComponent({
   name: 'BundleProductSelector',
