+++ packages/theme/nuxt.config.js
@@ -101,6 +101,7 @@ export default () => {
       '~/modules/catalog',
       '~/modules/customer',
       '~/modules/wishlist',
+      '~/modules/cart',
       '~/modules/checkout',
       '~/modules/review',
       ['nuxt-i18n', {
